+++ core/stk_set_optimizable_model_parameters.m
@@ -36,7 +36,7 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function model = stk_set_optimizable_model_parameters (model, value, select)
+function model = stk_set_optimizable_model_parameters (model, value)
 
 % Just a quick check
 if ~ (isstruct (model) && isfield (model, 'param'))
@@ -54,54 +54,23 @@ noiseparam_size = length (noiseparam);
 
 total_size = covparam_size + noiseparam_size;
 
-if nargin >= 3
-    
-    % 'select' must be a logical vector of appropriate length
-    % FIXME: support other types of selectors (such as indices) ?
-    if (~ islogical (select)) && (numel (select) ~= total_size)
-        stk_error (['Argument ''select'' must be a logical vector whose length is ' ...
-            'equal to the number of parameters.'], 'InvalidArgument');
-    end
-    
-    % Split selector
-    covparam_select = select(1:covparam_size);
-    noiseparam_select = select(covparam_size + (1:noiseparam_size));
-    
-    % Recompute block sizes
-    covparam_size = sum (covparam_select);
-    noiseparam_size = sum (noiseparam_select);
-    total_size = covparam_size + noiseparam_size;
-    
-end
-
 % Check length of value
 if numel (value) ~= total_size
     stk_error (['The length of ''value'' must be equal to ' ...
         'the number of parameters.'], 'IncorrectSize');
 end
-   
+
 % Change covariance parameter
 if covparam_size > 0
     covparam = value(1:covparam_size);
-    if nargin < 3
-        model.param = stk_set_optimizable_parameters ...
-            (model.param, covparam);
-    else
-        model.param = stk_set_optimizable_parameters ...
-            (model.param, covparam, covparam_select);
-    end
+    model.param = stk_set_optimizable_parameters (model.param, covparam);
 end
 
 % Change noise variance parameter
 if noiseparam_size > 0
     noiseparam = value(covparam_size + (1:noiseparam_size));
-    if nargin < 3
-        model.lognoisevariance = stk_set_optimizable_parameters ...
-            (model.lognoisevariance, noiseparam);
-    else
-        model.lognoisevariance = stk_set_optimizable_parameters ...
-            (model.lognoisevariance, noiseparam, noiseparam_select);
-    end
+    model.lognoisevariance = stk_set_optimizable_parameters ...
+        (model.lognoisevariance, noiseparam);
 end
 
 end % function